Initial conversions of torch network into an nn.Module
