+++ src/utils.py
@@ -37,6 +37,7 @@ import blitnet as bn
 
 from metrics import recallAtK, createPR
 from timeit import default_timer
+from os import path
 
 
 def get_patches2D(patch_size,image_pad):
@@ -184,6 +185,407 @@ def setSpikeRates(data,ids,device,dims,test_true,numImgs,numMods,intensity,locRe
                 spike_rates = torch.concat((spike_rates,torch.unsqueeze(init_rates,0)),0)
      
      return spike_rates
+
+# Check if pre-trained network exists, prompt if retrain or run        
+def checkTrainTest(self):
+    prompt = "A network with these parameters exists, re-train network? (y/n):\n"
+    self.logger.info('')
+    if path.isdir(self.training_out):
+        retrain = input(prompt)
+    else:
+        retrain = 'y'
+    return retrain
+
+def initialize(self,condition):
+    
+    '''
+    Network startup and initialization
+    '''
+    self.logger.info('')
+    self.logger.info('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    self.logger.info(condition+' startup and initialization')    
+    self.logger.info('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    self.logger.info('')
+    self.logger.info('Loading '+condition+' images')
+    
+    if condition == 'testing':
+        self.test_true = True
+        del self.filteredNames[self.number_testing_images:len(self.filteredNames)]
+        self.epoch = 1 # Only run the network once
+        self.location_repeat = 1 # One location repeat for testing
+        imgNum = self.number_testing_images
+    else:
+        imgNum = self.number_training_images
+        
+    # load the training images
+    self.imgs[condition], self.ids[condition] = ut.loadImages(self.test_true,
+                                        self.fullTrainPaths,
+                                        self.filteredNames,
+                                        [self.imWidth,self.imHeight],
+                                        self.num_patches,
+                                        self.testPath,
+                                        self.test_location)
+
+    self.spike_rates[condition] = ut.setSpikeRates(self.imgs[condition],
+                                       self.ids[condition],
+                                        self.device,
+                                        [self.imWidth,self.imHeight],
+                                        self.test_true,
+                                        imgNum,
+                                        self.number_modules,
+                                        self.intensity,
+                                        self.location_repeat)
+def train(self):
+
+    # remove contents of the weights folder
+    ut.clear_weights(self.training_out)
+
+    # create a new blitnet netowrk
+    self.logger.info('Creating network and setting weights')
+    net = bn.newNet(self.number_modules,self.imWidth*self.imHeight)
+    
+    # add the input layer
+    bn.addLayer(net,[self.number_modules,1,self.input_layer],
+                         0.0,0.0,0.0,0.0,0.0,False)   
+    
+    # add the feature layer
+    bn.addLayer(net,[self.number_modules,1,self.feature_layer],
+                    [0,self.theta_max],
+                    [self.f_rate[0],self.f_rate[1]],
+                     self.n_itp,
+                    [0,self.c],
+                     0,
+                     False)
+    
+    # sequentially set the feature firing rates for the feature layer
+    fstep = (self.f_rate[1]-self.f_rate[0])/self.feature_layer
+    
+    # loop through all modules and feature layer neurons
+    for x in range(self.number_modules):
+        for i in range(self.feature_layer):
+            net['fire_rate'][1][x][:,i] = self.f_rate[0]+fstep*(i+1)
+        
+    # add excitatory inhibitory connections for input and feature layer
+    bn.addWeights(net,0,1,[-1,0,1],[self.p_exc,self.p_inh],self.n_init)
+    self.init_weights = [net['W'][0].clone().detach(),net['W'][1].clone().detach()]
+
+    '''
+    Feature layer training
+    '''
+    self.logger.info('')
+    self.logger.info('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    self.logger.info('Training the input to feature layer')    
+    self.logger.info('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    self.logger.info('')
+    
+    self.logger.info('Setting spike rates from loaded images')
+    
+    # begin timer for network training
+    start = timeit.default_timer()
+    
+    # Set the spikes times for the input images
+    net['set_spks'][0] = torch.clone(self.spike_rates['training'])
+    self.spike_rates['training'] = []
+    if self.device == 'cuda':
+        torch.cuda.empty_cache()
+        gc.collect()    
+    layers = [len(net['W'])-2, len(net['W'])-1, len(net['W_lyr'])-1]
+    
+    # Train the input to feature layer for specified amount of epochs
+    for epoch in range(self.epoch):
+        net['step_num'] = 0
+        epochStart = timeit.default_timer()
+        
+        # loop through each image and train the network
+        for t in range(int(self.T)):
+            bn.runSim(net,1,self.device,layers)
+            # anneal learning rates
+            if np.mod(t,10)==0:
+                pt = pow(float(self.T-t)/self.T,self.annl_pow)
+                net['eta_ip'][1] = self.n_itp*pt
+                net['eta_stdp'][0] = self.n_init*pt
+                net['eta_stdp'][1] = -1*self.n_init*pt
+        
+        # print training details
+        self.logger.info('Epoch '+str(epoch+1)+' trained in: '
+              +str(round(timeit.default_timer()-epochStart,2))+'s')
+        self.logger.info('')
+        
+    self.logger.info('Finished training input to feature layer')
+    
+    '''
+    Preparations for feature to output layer training
+    '''
+    
+    # Turn off learning between input and feature layer
+    net['eta_ip'][1] = 0.0
+    if self.p_exc > 0.0: net['eta_stdp'][0] = 0.0
+    if self.p_inh > 0.0: net['eta_stdp'][1] = 0.0
+    
+    self.logger.info('Getting feature layer spikes for output layer training')
+
+    # get the feature spikes for training the output layer
+    net['x_feat'] = []
+    net['step_num'] = 0
+    for t in range(int(self.T)): # run blitnet without learning to get feature spikes
+        bn.runSim(net,1,self.device,layers)
+        net['x_feat'].append(net['x'][1]) # dictionary output of feature spikes
+        
+    # delete input spikes
+    net['set_spks'][0] = []
+    if self.device == 'cuda':
+        torch.cuda.empty_cache()
+        gc.collect()
+    
+    self.logger.info('Creating output layer')    
+    # Create and train the output layer with the feature layer
+    bn.addLayer(net,[self.number_modules,1,self.output_layer],
+                0.0,0.0,0.0,0.0,0.0,False)
+
+    # Add excitatory and inhibitory connections
+    bn.addWeights(net,1,2,[-1.0,0.0,1.0],[1.0,1.0],self.n_init)
+    self.init_weights.append(net['W'][2].clone().detach())
+    self.init_weights.append(net['W'][3].clone().detach())
+    
+    # Output spikes for spike forcing (final layer)
+    out_spks = torch.tensor([0],device=self.device,dtype=float)
+
+    '''
+    Output layer training
+    '''
+    self.logger.info('')
+    self.logger.info('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    self.logger.info('Training the feature to output layer')    
+    self.logger.info('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    self.logger.info('')
+    
+    net['spike_dims'] = 1 # change spike dims for output spike indexing
+    net['set_spks'][0] = [] # remove input spikes
+    layers = [len(net['W'])-2, len(net['W'])-1, len(net['W_lyr'])-1]
+    
+    # Train the feature to output layer for specified number of epochs 
+    for epoch in range(self.epoch):
+        net['step_num'] = 0
+        epochStart = timeit.default_timer()
+        
+        # Loop through all the spikes generated in the feature layer to train output
+        for t in range(self.T):
+            out_spks = torch.tensor([0],device=self.device,dtype=float)
+
+            net['set_spks'][-1] = torch.tile(out_spks,
+                                             (self.number_modules,
+                                              1,
+                                              self.output_layer))
+            net['x'][1] = net['x_feat'][t]
+            bn.runSim(net,1,self.device,layers)
+            # Anneal learning rates
+            if np.mod(t,10)==0:
+                pt = pow(float(self.T-t)/(self.T),self.annl_pow)
+                net['eta_ip'][2] = self.n_itp*pt
+                net['eta_stdp'][2] = self.n_init*pt
+                net['eta_stdp'][3] = -1*self.n_init*pt
+            if np.mod((t+1),(int(self.T/self.location_repeat))) == 0:
+                net['step_num'] = 0     
+        
+        # print training details
+        self.logger.info('Epoch '+str(epoch+1)+' trained in: '
+              +str(round(timeit.default_timer()-epochStart,2))+'s')
+        self.logger.info('')
+        
+    self.logger.info('Finished training feature to output layer')
+    self.logger.info('')
+    self.logger.info('~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    self.logger.info('Network trained in '+str(round(timeit.default_timer()-start,2))
+                                          +'s')
+    self.logger.info('~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    self.logger.info('')
+    
+    # Turn off learning
+    net['eta_ip'][2] = 0.0
+    net['eta_stdp'][2] = 0.0
+    net['eta_stdp'][3] = 0.0
+
+    # Clear the network output spikes
+    net['set_spks'][-1] = []
+    net['spike_dims'] = self.input_layer
+    
+    # Reset network details
+    net['sspk_idx'] = [0,0,0]
+    net['step_num'] = 0
+    net['spikes'] = [[],[],[]]
+    net['x'] = [[],[],[]]
+    net['x_feat'] = []
+    
+        
+    self.logger.info('Network formatting and saving...') 
+    
+    # Output the trained network
+    outputPkl = self.training_out + 'net.pkl'
+    with open(outputPkl, 'wb') as f:
+        pickle.dump(net, f)
+        
+    # output the ground truth image names for later testing
+    outputPkl = self.training_out + 'GT_imgnames.pkl'
+    with open(outputPkl, 'wb') as f:
+        pickle.dump(self.filteredNames, f)
+    
+    self.logger.info('Network succesfully saved!')
+    
+    # if using cuda, clear and dump the memory usage
+    if self.device =='cuda':
+        del net
+        del self.spike_rates
+        del self.imgs
+        torch.cuda.empty_cache()
+        gc.collect()
+       
+'''
+ Run the testing network
+'''
+def networktester(self):
+     
+    '''
+    Network startup and initialization
+    '''
+
+    # unpickle the network
+    self.logger.info('Unpickling the network')
+    with open(self.training_out+'net.pkl', 'rb') as f:
+         net = pickle.load(f)
+    
+    self.logger.info('Setting spike rates from loaded images')
+    # calculate input spikes from training images
+    
+    net['set_spks'][0] = self.spike_rates['testing']
+    
+    # unpickle the ground truth image names
+    with open(self.training_out+'GT_imgnames.pkl', 'rb') as f:
+         GT_imgnames = pickle.load(f)
+    
+    self.logger.info('')
+    self.logger.info('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    self.logger.info('Running test network')    
+    self.logger.info('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    self.logger.info('')
+    
+    # set number of correct places to 0
+    numcorrect = 0
+    net['spike_dims'] = self.input_layer
+    numpconc = []
+    start = timeit.default_timer()
+    for t in range(self.number_testing_images):
+        tonump = np.array([])
+        bn.testSim(net,device=self.device)
+        # output the index of highest amplitude spike
+        
+        tonump = np.append(tonump,np.reshape(net['x'][-1].cpu().numpy(),
+                                   [1,1,int(self.number_training_images)]))
+        
+        # detect if no output
+        if np.all(tonump == 0):
+            # find the strongest sub-threshold input index
+            nidx = np.argmax(torch.sub(net['x_input'][-1][0,0,:],
+                                  net['thr'][-1][0,0,:]).cpu().numpy())
+            tonump[nidx] = 0.5
+        else:
+            # find the highest output spike
+            nidx = np.argmax(tonump)
+        
+        gt_ind = GT_imgnames.index(self.filteredNames[t])
+
+        if gt_ind == nidx:
+            numcorrect += 1
+        
+        if self.validation: # get similarity matrix for PR curve generation
+           numpconc.append(tonump.tolist())
+    
+    self.p100r = round((numcorrect/self.number_testing_images)*100,2)
+    self.logger.info('Number of correct matches P@100R - '+str(self.p100r)+'%')
+
+    end = timeit.default_timer()
+    queryHertz = self.number_testing_images/(end-start)
+    self.logger.info('System queried at '+str(round(queryHertz,2))+'Hz')
+    
+    # if self.validation = True, get PR information and plot similarity matrix
+    if self.validation:
+        val.match_metrics(numpconc, 
+                          self.output_folder, 
+                          self.number_testing_images, 
+                          self.number_training_images, 
+                          self.logger)
+
+    # if using cuda, clear and dump the memory usage
+    if self.device =='cuda':
+        gc.collect()
+        torch.cuda.empty_cache()
+
+def sad(self):
+
+    print('')
+    print('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    print('Setting up Sum of Absolute Differences (SAD) calculations')    
+    print('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~')
+    print('')
+
+    sadcorrect = 0
+
+    # load the training images
+    self.location_repeat = 1 # switch to only do SAD on one dataset traversal
+    self.fullTrainPaths = self.fullTrainPaths[1]
+    self.test_true = False # testing images preloaded, load the training ones
+    # load the training images
+    self.imgs['training'], self.ids['training'] = ut.loadImages(self.test_true,
+                                        self.fullTrainPaths,
+                                        self.filteredNames,
+                                        [self.imWidth,self.imHeight],
+                                        self.num_patches,
+                                        self.testPath,
+                                        self.test_location)
+
+    # create database tensor
+    for ndx, n in enumerate(self.imgs['training']):
+        if ndx == 0:
+            db = torch.unsqueeze(n,0)
+        else:
+            db = torch.concat((db,torch.unsqueeze(n,0)),0)
+
+    def calc_sad(query, database, const):
+
+        SAD = torch.sum(torch.abs(torch.sub((database * const), (query * const))),
+                        (1,2),keepdim=True)
+        for n in range(2):
+            SAD = torch.squeeze(SAD,-1)
+        return SAD
+
+    # calculate SAD for each image to database and count correct number
+    imgred = 1/(self.imWidth*self.imHeight)
+    sad_concat = []
+    print('Running SAD')
+
+    start = timeit.default_timer()
+    for n, q in enumerate(self.imgs['testing']):
+        pixels = torch.empty([])
+         # create 3D tensor of query images
+        for o in range(self.number_testing_images):
+            if o == 0:
+                pixels = torch.unsqueeze(q,0)
+            else:
+                pixels = torch.concat((pixels,torch.unsqueeze(q,0)),0)
+
+        sad_score = calc_sad(pixels, db, imgred)
+
+        best_match = np.argmin(sad_score.cpu().numpy())
+        if n == best_match:
+            sadcorrect += 1
+
+    end = timeit.default_timer()   
+    p100r = round((sadcorrect/self.number_testing_images)*100,2)
+    print('')
+    print('Sum of absolute differences P@1: '+
+          str(p100r)+'%')
+    print('Sum of absolute differences queried at '
+          +str(round(self.number_testing_images/(end-start),2))+'Hz')
+
 
 # plot similarity matrices
 def plot_similarity(mat, name, cmap, ax=None, dpi=600):
